Move common logic to separate package, move constants to config
