--- Data Fetch (ZKTLS)/tls-notary/ZK/src/config.ts
@@ -1,5 +1,0 @@
-export const config = {
-  apiBase: process.env.API_BASE || 'http://localhost:3002/tunnels',
-};
-
-
