--- Data Fetch (ZKTLS)/tls-notary/ZK/src/types/dto.ts
@@ -1,15 +1,0 @@
-export interface TunnelCreateRequest {
-  localPort: number;
-  remoteHost: string;
-  remotePort: number;
-}
-
-
-export interface TunnelCreateResponse {
-  id?: number;
-  localPort: number;
-  remoteHost: string;
-  remotePort: number;
-  websocketProxyUrl: string;
-  pid?: number;
-}
