+++ Data Fetch (ZKTLS)/tls-notary/browser-extension/.env
@@ -1,0 +1,6 @@
+FRONTEND_PORT=8092
+API_PORT=3002
+API_BASE=http://localhost:3002/tunnels
+OPENBANKING_API=https://openbanking-api-826260723607.europe-west3.run.app/users/aaa/credit-score
+TLS_REMOTE_PORT=443
+TLS_LOCAL_PORT=8080
