+++ Data Fetch (ZKTLS)/tls-notary/browser-extension/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "http-parser-js": "^0.5.10",
     "buffer": "^6.0.3",
-    "nanoid": "^4.0.2"
+    "nanoid": "^4.0.2",
+    "tls-notary-shared": "file:../shared"
   }
 }
