--- Data Fetch (ZKTLS)/tls-notary/browser-extension/src/config.ts
@@ -1,5 +1,0 @@
-export const config = {
-  apiBase: 'http://localhost:3002/tunnels',
-};
-
-
