--- Data Fetch (ZKTLS)/tls-notary/browser-extension/src/script/worker.ts
@@ -1,9 +1,0 @@
-import * as Comlink from 'comlink';
-import init, { Prover, Attestation, Presentation } from 'tlsn-js';
-
-Comlink.expose({
-    init,
-    Prover,
-    Presentation,
-    Attestation,
-});
