+++ Data Fetch (ZKTLS)/tls-notary/shared/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "tls-notary-shared",
+  "version": "1.0.0",
+  "description": "Shared code for TLS Notary projects",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "browser": {
+    "process": "process/browser"
+  },
+  "scripts": {
+    "build": "tsc && cp src/script/worker.ts dist/script/",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "buffer": "^6.0.3",
+    "comlink": "^4.4.2",
+    "http-parser-js": "^0.5.10",
+    "nanoid": "^4.0.2",
+    "process": "^0.11.10",
+    "tlsn-js": "^0.1.0-alpha.10",
+    "@types/node": "^22.15.24"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.0"
+  }
+}
