+++ Data Fetch (ZKTLS)/tls-notary/shared/src/config.ts
@@ -1,0 +1,6 @@
+export const config = {
+  apiBase: process.env.API_BASE || 'http://localhost:3002/tunnels',
+  openbankingApi: process.env.OPENBANKING_API || 'https://openbanking-api-826260723607.europe-west3.run.app',
+  tlsRemotePort: process.env.TLS_REMOTE_PORT || '443',
+  tlsLocalPort: process.env.TLS_LOCAL_PORT || '8080'
+};
