+++ Data Fetch (ZKTLS)/tls-notary/shared/src/index.ts
@@ -1,0 +1,20 @@
+// Export all shared code
+
+// Services
+export * from './services/ITLSNotaryService';
+export * from './services/MockTLSNotaryService';
+export * from './services/TunnelService';
+
+// Types
+export * from './types/dto';
+export * from './types/tls';
+
+// Scripts
+export * from './script/generateProofs';
+export * from './script/worker';
+
+// Utils
+export * from './utils/di';
+
+// Config
+export * from './config';
