+++ Data Fetch (ZKTLS)/tls-notary/shared/src/script/worker.ts
