+++ Data Fetch (ZKTLS)/tls-notary/shared/src/types/dto.ts
