+++ Data Fetch (ZKTLS)/tls-notary/shared/src/types/tls.ts
